# wb-landing-builder/docs/swagger.yaml
basePath: /
definitions:
  auth.ErrorResponse:
    properties:
      error:
        description: Error описание произошедшей ошибки.
        example: описание ошибки...
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        description: |-
          Email адрес зарегистрированного пользователя.
          Required: true
        example: newuser@example.com
        type: string
      password:
        description: |-
          Пароль пользователя.
          Required: true
        example: SuperSecretPass123
        type: string
    type: object
  auth.Me:
    properties:
      email:
        description: Email адрес электронной почты пользователя.
        example: newuser@example.com
        type: string
      id:
        description: ID уникальный идентификатор пользователя.
        example: 507f191e810c19729de860ea
        type: string
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        description: |-
          RefreshToken токен, полученный при предыдущем входе или обновлении.
          Required: true
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        description: |-
          Email адрес для регистрации.
          Required: true
        example: newuser@example.com
        type: string
      password:
        description: |-
          Пароль для доступа к аккаунту.
          Required: true
        example: SuperSecretPass123
        type: string
    type: object
  auth.TokenResponse:
    properties:
      access_token:
        description: AccessToken токен для авторизации защищенных запросов.
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        description: ExpiresIn время жизни access токена в секундах.
        example: 3600
        type: integer
      refresh_token:
        description: RefreshToken токен для получения новой пары access/refresh.
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  publishing.ErrorResponse:
    properties:
      error:
        example: описание ошибки...
        type: string
    type: object
  publishing.Publication:
    properties:
      assets_path:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      project_id:
        type: string
      status:
        $ref: '#/definitions/publishing.PublicationStatus'
      version:
        type: integer
    type: object
  publishing.PublicationIDsResponse:
    properties:
      ids:
        example:
        - 550e8400-e29b-41d4-a716-446655440000
        items:
          type: string
        type: array
    type: object
  publishing.PublicationStatus:
    enum:
    - PENDING
    - PROCESSING
    - FINISHED
    - FAILED
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusProcessing
    - StatusFinished
    - StatusFailed
  storage.ErrorResponse:
    properties:
      error:
        description: Error описание произошедшей ошибки.
        example: описание ошибки...
        type: string
    type: object
  storage.Mutation:
    type: object
  storage.OperationType:
    enum:
    - create
    - update
    - delete
    type: string
    x-enum-varnames:
    - OperationCreate
    - OperationUpdate
    - OperationDelete
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления черновиками лендингов и аутентификации пользователей.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WB Landing Builder API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает токены доступа.
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токены доступа
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Вход
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      consumes:
      - application/json
      description: Возвращает ID и email авторизованного пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: ID и email пользователя
          schema:
            $ref: '#/definitions/auth.Me'
        "401":
          description: Не авторизован или пользователь не найден
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Мой профиль
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет пару токенов, используя refresh токен.
      parameters:
      - description: Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новая пара токенов
          schema:
            $ref: '#/definitions/auth.TokenResponse'
        "400":
          description: Отсутствует refresh токен
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Невалидный refresh токен
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Обновление токенов
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает его ID и email.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ID и email созданного пользователя
          schema:
            $ref: '#/definitions/auth.Me'
        "400":
          description: Ошибка валидации или формата
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "409":
          description: Пользователь уже существует
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Регистрация
      tags:
      - Auth
  /api/v1/storage/{project_id}:
    get:
      consumes:
      - application/json
      description: Возвращает актуальную версию страницы для указанного проекта.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JSON контент страницы
          schema:
            $ref: '#/definitions/storage.Mutation'
        "400":
          description: Отсутствует project_id
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
        "500":
          description: Ошибка получения данных
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить последний черновик
      tags:
      - Storage
  /api/v1/storage/{project_id}/mutations:
    post:
      consumes:
      - application/json
      description: Применяет JSON-мутацию к указанному проекту после проверки схемы
        и лимитов.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      - description: Объект мутации
        in: body
        name: mutation
        required: true
        schema:
          $ref: '#/definitions/storage.Mutation'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное применение, возвращает версию
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
        "400":
          description: Ошибка валидации или неверный запрос
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
        "413":
          description: Превышен размер payload
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
        "429":
          description: Превышен лимит запросов
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Применить мутацию
      tags:
      - Storage
  /api/v1/storage/{project_id}/publications:
    get:
      description: Возвращает ID всех публикаций для указанного проекта (от новых
        к старым). Доступ только владельцу проекта.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список ID публикаций
          schema:
            $ref: '#/definitions/publishing.PublicationIDsResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "403":
          description: Нет доступа к проекту
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "404":
          description: Проект не найден
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список ID публикаций проекта
      tags:
      - Publications
    post:
      description: Создаёт запись публикации со статусом PENDING и ставит задачу на
        рендер в очередь. Доступ только владельцу проекта.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Публикация поставлена в очередь (status=PENDING)
          schema:
            $ref: '#/definitions/publishing.Publication'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "403":
          description: Нет доступа к проекту
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "404":
          description: Черновик не найден
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать публикацию
      tags:
      - Publications
  /api/v1/storage/{project_id}/publications/{id}:
    delete:
      description: Удаляет публикацию и связанные файлы. Доступ только владельцу проекта.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      - description: ID публикации
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Публикация удалена
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "403":
          description: Нет доступа к проекту
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "404":
          description: Публикация или проект не найдены
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить публикацию
      tags:
      - Publications
    get:
      description: Возвращает метаданные публикации и текущий статус (PENDING, PROCESSING,
        FINISHED, FAILED). Доступ только владельцу проекта.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      - description: ID публикации
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Метаданные публикации
          schema:
            $ref: '#/definitions/publishing.Publication'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "403":
          description: Нет доступа к проекту
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "404":
          description: Публикация или проект не найдены
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/publishing.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить публикацию
      tags:
      - Publications
  /api/v1/storage/{project_id}/versions/{version}:
    get:
      consumes:
      - application/json
      description: Возвращает страницу указанной версии для проекта.
      parameters:
      - description: ID проекта
        in: path
        name: project_id
        required: true
        type: string
      - description: Номер версии
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: JSON контент страницы
          schema:
            type: string
        "400":
          description: Неверный ID или версия
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
        "500":
          description: Ошибка получения данных
          schema:
            $ref: '#/definitions/storage.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить версию черновика
      tags:
      - Storage
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"